+++ platform/qt/src/qmapboxgl.cpp
@@ -626,8 +626,12 @@ void QMapboxGL::addAnnotationIcon(const QString &name, const QImage &sprite)
     auto img = std::make_unique<uint8_t[]>(swapped.byteCount());
     memcpy(img.get(), swapped.constBits(), swapped.byteCount());
 
-    d_ptr->mapObj->addAnnotationIcon(name.toStdString(), std::make_shared<mbgl::SpriteImage>(
-        mbgl::PremultipliedImage { size_t(swapped.width()), size_t(swapped.height()), std::move(img) }, 1.0));
+    d_ptr->mapObj->addAnnotationIcon(
+        name.toStdString(),
+        std::make_shared<mbgl::SpriteImage>(
+            mbgl::PremultipliedImage{ static_cast<uint16_t>(swapped.width()),
+                                      static_cast<uint16_t>(swapped.height()), std::move(img) },
+            1.0));
 }
 
 QPointF QMapboxGL::pixelForCoordinate(const Coordinate &coordinate_) const
