+++ src/mbgl/util/raster.hpp
@@ -38,12 +38,12 @@ public:
     GLuint getID() const;
 
     // size of uploaded image.
-    std::array<size_t, 2> getSize() const;
+    std::array<uint16_t, 2> getSize() const;
 
 private:
     // raw pixels have been loaded
     std::atomic<bool> loaded { false };
-    std::array<size_t, 2> size = {{ 0, 0 }};
+    std::array<uint16_t, 2> size = {{ 0, 0 }};
 
     // filters
     Scaling filter = Scaling::Nearest;
